Parser/requirements.txt
antlr4-python3-runtime==4.9.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
cmake==3.27.2
comet-ml==3.33.9
configobj==5.0.8
contourpy==1.1.0
cycler==0.11.0
diskcache==5.6.1
dulwich==0.21.5
everett==3.1.0
filelock==3.12.2
fonttools==4.42.1
idna==3.4
imageio==2.31.1
imgaug==0.4.0
importlib-resources==6.0.1
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
lazy-loader==0.3
lit==16.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pkgutil-resolve-name==1.3.10
psutil==5.9.5
Pygments==2.16.1
pyntcloud==0.3.1
pyparsing==3.0.9
python-box==6.1.0
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.5.2
rpds-py==0.9.2
scikit-image==0.21.0
scipy==1.10.1
semantic-version==2.10.0
sentry-sdk==1.29.2
shapely==2.0.1
simplejson==3.19.1
six==1.16.0
sympy==1.12
thop==0.1.1.post2209072238
tifffile==2023.7.10
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
websocket-client==1.3.3
wrapt==1.15.0
wurlitzer==3.0.3
xformers==0.0.21
zipp==3.16.2
icecream==2.1.3